pyproject.toml
[tool.poetry]
name = "apis-instance-sicprod-dev-test"
version = "0.1.0"
description = "APIS instance SiCProD test"
authors = ["Birger Schacht <birger.schacht@oeaw.ac.at>"]
license = "MIT"
packages = [{include = "apis_ontology"}]


[tool.poetry.dependencies]
python = "^3.11"
django = ">=4.1,<4.2"
mysqlclient = "^2.0.3"
apis-core = { git = "https://github.com/acdh-oeaw/apis-core-rdf.git", branch = "sicprod-dev" }
webpage = { git = "https://github.com/acdh-oeaw/apis-webpage.git", tag = "v0.2.0" }
apis-bibsonomy = { git = "https://github.com/acdh-oeaw/apis-bibsonomy.git", branch = "birger/fix-referenceonform-issues" }
apis-acdhch-default-settings = { git = "https://github.com/acdh-oeaw/apis-acdhch-default-settings.git", tag = "v0.1.7" }
django-versions = { git = "https://github.com/b1rger/django-versions.git", tag = "v0.2.0" }

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
